#2070: Moved all nontrivial controllers to new file.
